Refactor QG diagnostics to shared class
