+++ pyqg/layered_model.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numpy import pi
-from . import model
+from . import qg_diagnostics
 
 try:
     import mkl
@@ -14,7 +14,7 @@ try:
 except ImportError:
     pass
 
-class LayeredModel(model.Model):
+class LayeredModel(qg_diagnostics.QGDiagnostics):
     r"""Layered quasigeostrophic model.
 
     This model is meant to represent flows driven by baroclinic instabilty of a
@@ -239,28 +239,6 @@ class LayeredModel(model.Model):
         # self.filtr[wvx<=cphi] = 1.
 
     ### All the diagnostic stuff follows. ###
-    def _calc_cfl(self):
-        return np.abs(
-            np.hstack([self.u + self.Ubg[:,np.newaxis,np.newaxis], self.v])
-        ).max()*self.dt/self.dx
-
-    # calculate KE: this has units of m^2 s^{-2}
-    #   (should also multiply by H1 and H2...)
-    def _calc_ke(self):
-        ke = 0.
-        for j in range(self.nz):
-            ke += .5*self.Hi[j]*self.spec_var(self.wv*self.ph[j])
-        return ke.sum() / self.H
-
-    # calculate eddy turn over time
-    # (perhaps should change to fraction of year...)
-    def _calc_eddy_time(self):
-        """ estimate the eddy turn-over time in days """
-        ens = 0.
-        for j in range(self.nz):
-            ens = .5*self.Hi[j] * self.spec_var(self.wv2*self.ph[j])
-
-        return 2.*pi*np.sqrt( self.H / ens.sum() ) / 86400
 
     def _calc_derived_fields(self):
 
@@ -279,13 +257,7 @@ class LayeredModel(model.Model):
     def _initialize_model_diagnostics(self):
         """ Extra diagnostics for layered model """
 
-        self.add_diagnostic('entspec',
-                description='barotropic enstrophy spectrum',
-                function= (lambda self:
-                    np.abs((self.Hi[:,np.newaxis,np.newaxis]*self.qh).sum(axis=0))**2/self.H/self.M**2),
-                units='m s^-2',
-                dims=('l','k')
-        )
+        super()._initialize_model_diagnostics()
 
         self.add_diagnostic('KEspec_modal',
                 description='modal kinetic energy spectra',
@@ -328,38 +300,5 @@ class LayeredModel(model.Model):
                 dims=('l','k')
         )
         
-        self.add_diagnostic('APEgenspec',
-                    description='the spectrum of the rate of generation of available potential energy',
-                    function =(lambda self: (self.Hi[:,np.newaxis,np.newaxis]*
-                                (self.Ubg[:,np.newaxis,np.newaxis]*self.k +
-                                 self.Vbg[:,np.newaxis,np.newaxis]*self.l)*
-                                (1j*self.ph.conj()*self.Sph).real).sum(axis=0)/self.H/self.M**2),
-                units='m^2 s^-3',
-                dims=('l','k')
-        )
 
-        self.add_diagnostic('ENSflux',
-                 description='barotropic enstrophy flux',
-                 function = (lambda self: (-self.Hi[:,np.newaxis,np.newaxis]*
-                              (self.qh.conj()*self.Jq).real).sum(axis=0)/self.H/self.M**2),
-                units='s^-3',
-                dims=('l','k')
-        )
-
-        self.add_diagnostic('ENSgenspec',
-                    description='the spectrum of the rate of generation of barotropic enstrophy',
-                    function = (lambda self:
-                                (self.Hi[:,np.newaxis,np.newaxis]*((self.ilQx-self.ikQy)*
-                                self.Sph.conj()*self.ph).real).sum(axis=0)/self.H/self.M**2),
-                units='s^-3',
-                dims=('l','k')
-        )
-
-        self.add_diagnostic('ENSfrictionspec',
-                    description='the spectrum of the rate of dissipation of barotropic enstrophy due to bottom friction',
-                    function = (lambda self: self.rek*self.Hi[-1]/self.H*self.wv2*
-                        (self.qh[-1].conj()*self.ph[-1]).real/self.M**2), 
-                units='s^-3',
-                dims=('l','k')
-        )
