+++ pyqg/tests/test_xarray_output.py
@@ -40,8 +40,8 @@ expected_diags = [
     'ENSflux',
     'ENSgenspec',
     'paramspec',
-    'paramspec_apeflux',
-    'paramspec_keflux',
+    'paramspec_APEflux',
+    'paramspec_KEflux',
     'Dissspec',
     'KEfrictionspec',
     'ENSDissspec',
